chore: fix pre-existing lint and configure mypy
Clean up 20 ruff findings across the repo (unused imports, unsorted
import blocks, a redundant int() cast, an if/else simplified to a
default + if). Add a [tool.mypy] config: the pydantic plugin (so
Settings() with no args is valid) and ignore_missing_imports for
boto3/botocore/sigmf (no upstream stubs). Repo is now ruff- and
mypy-clean, all 71 tests green.

+++ src/aerolake/quality/metrics.py
@@ -30,7 +30,6 @@ from __future__ import annotations
 
 import numpy as np
 
-
 # ---------------------------------------------------------------------------
 # Constants
 # ---------------------------------------------------------------------------
@@ -371,7 +370,8 @@ def compute_sample_completeness(
     # This is the contract: capture for D seconds at rate R should yield
     # exactly D * R samples. Any deviation is suspect.
     # We round to an integer because sample counts are always whole.
-    expected_count = int(round(expected_duration_s * sample_rate))
+    # round() with no second argument already returns an int, so no int() cast.
+    expected_count = round(expected_duration_s * sample_rate)
 
     # Step 2 — Guard against division by zero.
     # If somehow expected_count is 0 (would be a programming error,
